% Auto-generated: do not edit by hand
\name{divider}

\alias{divider}

\title{Divider component}

\description{
Divider visually separate contents with a thin line and margin on all sides.  Dividers work best in flex layouts where they will adapt to orientation without additional  styles. Otherwise, a divider will appear as a full-width 1px-high block element.
}

\usage{
divider(id=NULL, className=NULL, tagName=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{tagName}{Character. HTML tag to use for element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

