% Auto-generated: do not edit by hand
\name{drawer}

\alias{drawer}

\title{Drawer component}

\description{

}

\usage{
drawer(children=NULL, id=NULL, autoFocus=NULL, className=NULL,
isOpen=NULL, key=NULL, role=NULL, title=NULL, ...)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{autoFocus}{Logical. Whether the overlay should acquire application focus when it first opens.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{isOpen}{Logical. Toggles the visibility of the overlay and its children. This prop is required because the component is controlled.}

\item{key}{Character. A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}

\item{role}{Character. The ARIA role attribute}

\item{title}{Character. }


\item{...}{wildcards allowed have the form: `'aria-*', 'data-*'`}

}

\value{named list of JSON elements corresponding to React.js properties and their values}

