% Auto-generated: do not edit by hand
\name{form}

\alias{form}

\title{Form component}

\description{
Custom component that is a collection of FormGroups with submit button that may have custom validation to enable  https://blueprintjs.com/docs/#core/components/dialog.dialog  @param props @returns {*} @constructor
}

\usage{
form(children=NULL, id=NULL, key=NULL, ...)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{key}{Character. A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}


\item{...}{wildcards allowed have the form: `'aria-*'`}

}

\value{named list of JSON elements corresponding to React.js properties and their values}

