% Auto-generated: do not edit by hand
\name{h6}

\alias{h6}

\title{H6 component}

\description{

}

\usage{
h6(children=NULL, id=NULL, className=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{className}{Character. Often used with CSS to style elements with common properties.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

