% Auto-generated: do not edit by hand
\name{menu}

\alias{menu}

\title{Menu component}

\description{
Menus display lists of interactive items. @param {*} props
}

\usage{
menu(children=NULL, id=NULL, className=NULL, key=NULL,
large=NULL, role=NULL, ...)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{key}{Character. A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}

\item{large}{Logical. Whether the menu items in this menu should use a large appearance.
@default false}

\item{role}{Character. The ARIA role attribute}


\item{...}{wildcards allowed have the form: `'aria-*', 'data-*'`}

}

\value{named list of JSON elements corresponding to React.js properties and their values}

