% Auto-generated: do not edit by hand
\name{menuDivider}

\alias{menuDivider}

\title{MenuDivider component}

\description{
Use MenuDivider to separate menu sections. Optionally, add a title to the divider. @param {*} props
}

\usage{
menuDivider(id=NULL, className=NULL, title=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{title}{Character. Optional header title.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

