% Auto-generated: do not edit by hand
\name{navbarDivider}

\alias{navbarDivider}

\title{NavbarDivider component}

\description{

}

\usage{
navbarDivider(id=NULL, className=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

