% Auto-generated: do not edit by hand
\name{navbarGroup}

\alias{navbarGroup}

\title{NavbarGroup component}

\description{

}

\usage{
navbarGroup(children=NULL, id=NULL, align=NULL, className=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{align}{Character. The side of the navbar on which the group should appear. The Alignment enum provides constants for these values.
One of (`left`, `center`, `right`)}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

